More Linux/GCC compability, still missing some blanks.

+++ gui/controls/CValueComboBox.cpp
@@ -1,30 +1,30 @@
 /*************************************************************************************
- 
+
 	cpl - cross-platform library - v. 0.1.0.
- 
+
 	Copyright (C) 2016 Janus Lynggaard Thorborg (www.jthorborg.com)
- 
+
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU General Public License as published by
 	the Free Software Foundation, either version 3 of the License, or
 	(at your option) any later version.
- 
+
 	This program is distributed in the hope that it will be useful,
 	but WITHOUT ANY WARRANTY; without even the implied warranty of
 	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 	GNU General Public License for more details.
- 
+
 	You should have received a copy of the GNU General Public License
 	along with this program.  If not, see <http://www.gnu.org/licenses/>.
- 
+
 	See \licenses\ for additional details on licenses associated with this program.
- 
+
 **************************************************************************************
- 
+
 	file:CValueComboBox.cpp
- 
+
 		Source code for CValueComboBox.h
- 
+
 *************************************************************************************/
 
 #include "CValueComboBox.h"
@@ -108,7 +108,7 @@ namespace cpl
 		if (2 > size)
 			size = 2;
 		idx = cpl::Math::confineTo(idx, 1, size);
-		
+
 		return iCtrlPrec_t(idx - 1) / (size - 1);
 	}
 
@@ -178,7 +178,7 @@ namespace cpl
 
 	bool CValueComboBox::setEnabledStateFor(std::size_t entry, bool toggle)
 	{
-		if (entry < values.size() && entry != -1)
+		if (entry < values.size() && entry != static_cast<std::size_t>(-1))
 		{
 			bool selectNew = box.getSelectedId() == int(entry + 1) && !toggle;
 			box.setItemEnabled((int)entry + 1, !!toggle);
